Using async - await

+++ server/server.js
@@ -60,24 +60,22 @@ app.get('/todos/:id', authenticate, (req, res)=>{
 
 });
 
-app.delete('/todos/:id', authenticate, (req, res)=>{
+app.delete('/todos/:id', authenticate, async (req, res)=>{
   const id = req.params.id;
 
   if (!ObjectID.isValid(id)) {
     return res.status(404).send();
   }
 
-  Todo.findOneAndRemove({
-    _id: id,
-    _creator: req.user._id
-  }).then((todo)=>{
+  try {
+    const todo = await Todo.findOneAndRemove({_id: id, _creator: req.user._id});
     if (!todo) {
-      return res.status(404).send();
+      res.status(404).send();
     }
     res.status(200).send({todo});
-  }).catch((e)=>{
+  } catch (e) {
     res.status(400).send(e);
-  });
+  }
 });
 
 app.patch('/todos/:id', authenticate,  (req, res)=>{
@@ -109,40 +107,41 @@ app.patch('/todos/:id', authenticate,  (req, res)=>{
 
 });
 
-app.post('/users', (req, res) => {
+app.post('/users', async (req, res) => {
   const user = new User(_.pick(req.body, ['email', 'password']));
 
-  user.save().then((user) => {
-  return user.generateAuthToken();
-  }).then((token)=>{
+  try {
+    await user.save();
+    const token = await user.generateAuthToken();
     res.header('x-auth', token).status(200).send(user);
-  }).catch((e) => {
+  } catch (e) {
     res.status(400).send(e);
-  });
+  }
+
 });
 
-app.post('/users/login', (req, res) => {
-  var body = _.pick(req.body, ['email', 'password']);
-
-  User.findByCredentials(body.email, body.password).then((user) => {
-      return user.generateAuthToken().then((token) => {
-        res.header('x-auth', token).send(user);
-    });
-  }).catch((e) => {
-    res.status(400).send();
-  });
+app.post('/users/login', async (req, res) => {
+  try {
+    const body = _.pick(req.body, ['email', 'password']);
+    const user = await User.findByCredentials(body.email, body.password);
+    const token = await user.generateAuthToken();
+    res.header('x-auth', token).send(user);
+  } catch (e) {
+      res.status(400).send();
+  }
 });
 
 app.get('/users/me', authenticate, (req, res)=>{
   res.send(req.user);
 });
 
-app.delete('/users/me/token', authenticate, (req, res) => {
-  req.user.removeToken(req.token).then(() => {
-    res.status(200).send();
-  }, () => {
-    res.status(400).send();
-  })
+app.delete('/users/me/token', authenticate, async (req, res) => {
+  try {
+    await req.user.removeToken(req.token);
+    res.status(200).send()
+  } catch (e) {
+      res.status(400).send();
+  }
 });
 
 app.listen(PORT, () => console.log(`App running on port ${PORT}`));
